requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
openai==1.59.5
chromadb==0.5.23
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
PyPDF2==3.0.1
python-docx==1.1.2
PyMuPDF==1.25.1
httpx==0.28.1
structlog==24.4.0
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
